+++ lib/index.js
@@ -9,6 +9,8 @@ var query_1 = require("./query");
 exports.query = query_1.default;
 var params_1 = require("./params");
 exports.params = params_1.default;
+var logger_1 = require("./logger");
+exports.logger = logger_1.default;
 var route_1 = require("./route");
 exports.default = route_1.default;
 exports.route = route_1.default;
