+++ lib/logger.js
@@ -1,21 +1,36 @@
 "use strict";
-const { Signale } = require('signale');
-const defaults = {
-    color: 'green',
-    logLevel: 'info'
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
 };
-const options = {
-    types: {
-        request: Object.assign({}, defaults, { label: 'REQUEST' }),
+Object.defineProperty(exports, "__esModule", { value: true });
+const path_1 = __importDefault(require("path"));
+const logger_1 = __importDefault(require("./utils/logger"));
+const fs_1 = require("fs");
+let cwd = process.cwd();
+const requirePaths = [];
+while (cwd !== '/') {
+    requirePaths.push(cwd);
+    cwd = path_1.default.join(cwd, '../');
+}
+const configPaths = requirePaths.map((p) => path_1.default.join(p, 'light.config.js'));
+const config = configPaths.reduce((acc, val) => {
+    let conf = {};
+    if (fs_1.existsSync(val)) {
+        try {
+            console.log('requiring');
+            conf = require(val);
+        }
+        catch (err) {
+            logger_1.default.error(`unable to import light.config.js: ${err}`);
+        }
+    }
+    return Object.assign({}, acc, conf);
+}, {});
+exports.default = (message) => {
+    if (config.logger) {
+        console.log(`${config.logger} - ${message}`);
+    }
+    else {
+        console.log(message);
     }
 };
-['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS', 'TRACE', 'PATCH'].forEach((method) => {
-    options.types[method] = Object.assign({}, defaults, { label: method });
-});
-const signale = new Signale(options);
-signale.config({
-    displayTimestamp: true,
-    displayDate: false,
-    underlineLabel: false,
-});
-module.exports = signale;
