+++ lib/route.js
@@ -5,37 +5,12 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 const micro_1 = require("micro");
 const aws_serverless_micro_1 = __importDefault(require("aws-serverless-micro"));
-const path_1 = __importDefault(require("path"));
 const micro_boom_1 = require("micro-boom");
-const logger_1 = __importDefault(require("./utils/logger"));
-const fs_1 = require("fs");
 const { LIGHT_ENVIRONMENT } = process.env;
 const youchErrors = require('./utils/route/youch'); // eslint-disable-line @typescript-eslint/no-var-requires
 const logger = require('./utils/route/logger'); // eslint-disable-line @typescript-eslint/no-var-requires
-let cwd = process.cwd();
-const requirePaths = [];
-while (cwd !== '/') {
-    requirePaths.push(cwd);
-    cwd = path_1.default.join(cwd, '../');
-}
-const configPaths = requirePaths.map((p) => path_1.default.join(p, 'light.config.js'));
-const config = configPaths.reduce((acc, val) => {
-    let conf = {};
-    if (fs_1.existsSync(val)) {
-        try {
-            console.log('requiring');
-            conf = require(val);
-        }
-        catch (err) {
-            logger_1.default.error(`unable to import light.config.js: ${err}`);
-        }
-    }
-    return Object.assign({}, acc, conf);
-}, {});
 exports.default = (route) => {
-    console.log('handler config is', config);
     const proxy = async (Req, Res) => {
-        console.log('proxy config is', config);
         let exec = async (req, res) => {
             const middleware = route.middleware || [];
             for (const mw of middleware) { // eslint-disable-line
@@ -44,7 +19,6 @@ exports.default = (route) => {
                     return null;
                 }
             }
-            console.log('config is', config);
             return route.handler(req, res);
         };
         const plugins = route.plugins || [];
@@ -67,7 +41,6 @@ exports.default = (route) => {
     Object.assign(fn, route, { handler: fn });
     fn.log = true;
     fn.module = __dirname;
-    console.log('final config is', config);
     /* istanbul ignore if */
     if (isNetlify || isAWS) {
         return {
