+++ src/index.ts
@@ -6,6 +6,7 @@ exports = route.default;
 export { default as server } from './server';
 export { default as query } from './query';
 export { default as params } from './params';
+export { default as logger } from './logger';
 export { default, default as route, default as light } from './route';
 export {
   buffer,
